requirements.txt
aws-cdk-lib
constructs
cdklabs.generative_ai_cdk_constructs
